# api/templates/openapi.json
{
   "openapi": "3.1.0",
   "info": {
      "title": "biodata-registry-service",
      "description": "\n## biodata-registry-service\n\nService to fetch data from Registry.\n\n",
      "version": "0.0.0"
   },
   "paths": {
...
      "/data_asset": {
         "post": {
            "tags": [
               "core"
            ],
            "summary": "Create Data Asset",
            "operationId": "create_data_asset",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/DataAssetCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DataAssets"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Data Asset",
            "operationId": "get_data_asset",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DataAssets"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "core"
            ],
            "summary": "Delete",
            "operationId": "delete_data_asset",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "core"
            ],
            "summary": "Update",
            "operationId": "update_data_asset",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/DataAssetUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DataAssets"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/data_assets": {
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Data Assets",
            "operationId": "get_data_assets",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/DataAssets"
                           },
                           "title": "Response Get Data Assets"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/instrument": {
         "post": {
            "tags": [
               "core"
            ],
            "summary": "Create Instrument",
            "operationId": "create_instrument",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/InstrumentCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Instruments"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Instrument",
            "operationId": "get_instrument",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Instruments"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "core"
            ],
            "summary": "Delete",
            "operationId": "delete_instrument",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "core"
            ],
            "summary": "Update",
            "operationId": "update_instrument",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/InstrumentUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Instruments"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/instruments": {
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Instruments",
            "operationId": "get_instruments",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/Instruments"
                           },
                           "title": "Response Get Instruments"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "schemas": {
...
         "DataAssetCreate": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "location": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Location"
               },
               "external_links": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "External Links"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "required": [
               "name",
               "location"
            ],
            "title": "DataAssetCreate"
         },
         "DataAssetUpdate": {
            "properties": {
               "name": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 254
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Name"
               },
               "location": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 254
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Location"
               },
               "external_links": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "External Links"
               },
               "data": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "DataAssetUpdate"
         },
         "DataAssets": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "location": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Location"
               },
               "external_links": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "External Links"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               }
            },
            "type": "object",
            "required": [
               "name",
               "location"
            ],
            "title": "DataAssets"
         },
         "HTTPValidationError": {
            "properties": {
               "detail": {
                  "items": {
                     "$ref": "#/components/schemas/ValidationError"
                  },
                  "type": "array",
                  "title": "Detail"
               }
            },
            "type": "object",
            "title": "HTTPValidationError"
         },
         "InstrumentCreate": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "required": [
               "name"
            ],
            "title": "InstrumentCreate"
         },
         "InstrumentUpdate": {
            "properties": {
               "name": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 254
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Name"
               },
               "data": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "InstrumentUpdate"
         },
         "Instruments": {
            "properties": {
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               },
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "required": [
               "name"
            ],
            "title": "Instruments"
         },
...
         "ValidationError": {
            "properties": {
               "loc": {
                  "items": {
                     "anyOf": [
                        {
                           "type": "string"
                        },
                        {
                           "type": "integer"
                        }
                     ]
                  },
                  "type": "array",
                  "title": "Location"
               },
               "msg": {
                  "type": "string",
                  "title": "Message"
               },
               "type": {
                  "type": "string",
                  "title": "Error Type"
               },
               "input": {
                  "title": "Input"
               },
               "ctx": {
                  "type": "object",
                  "title": "Context"
               }
            },
            "type": "object",
            "required": [
               "loc",
               "msg",
               "type"
            ],
            "title": "ValidationError"
         }
      }
   }
}